requirements.txt
aerich==0.8.2
aiofiles==24.1.0
aiogram==3.19.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asyncclick==8.1.8.0
asyncpg==0.30.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
dictdiffer==0.9.0
docker==7.1.0
frozenlist==1.5.0
idna==3.10
iso8601==2.1.0
magic-filter==1.0.12
multidict==6.4.3
propcache==0.3.1
pydantic==2.10.6
pydantic_core==2.27.2
pypika-tortoise==0.5.0
pytz==2025.2
pywin32==310
redis==5.2.1
requests==2.32.3
sniffio==1.3.1
tortoise-orm==0.25.0
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.19.0
